requirements.txt
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
setuptools
dj-rest-auth==6.0.0
pytest-django==4.9.0
Django